Move network services tasks to sub-modules
